Refactored error message and organized files
