+++ FrontEnd/src/Pages/Home.js
@@ -1,11 +1,14 @@
 import React from 'react'
-import Subscribe from '../components/Subscribe/Subscribe'
-import SpeakerPartners from '../components/SpeakerPartners/SpeakerPartners'
+import Subscribe from '../components/SubscribeComponent/Subscribe'
+import SpeakerPartners from '../components/SpeakerPartnersComponent/SpeakerPartners'
+import Email from "../components/EmailComponent/Email";
 function Home(){
     return(
         <div>
             <SpeakerPartners />
-
+            <div style={{ display: "flex", justifyContent: "center" }}>
+                <Email />
+            </div>
         </div>
     );
 }
