+++ FrontEnd/src/components/EmailComponent/Email.tsx
@@ -16,6 +16,18 @@ const Email = () => {
 
   const [contactNumber, setContactNumber] = useState("000000");
   const [statusMessage, setStatusMessage] = useState("");
+  const errMsg = (err: any) => {return(
+    <div
+          role="alert"
+          style={{
+            display: "flex",
+            justifyContent: "center",
+            color: "red",
+          }}
+        >
+          {err.message}
+        </div>
+  )}
 
   React.useEffect(() => {
     if (isSubmitSuccessful) {
@@ -65,18 +77,8 @@ const Email = () => {
   return (
     <form id="contact-form" onSubmit={handleSubmit(onSubmit)}>
       <input type="hidden" name="id" value={contactNumber} />
-      {errors.yourName && (
-        <div
-          role="alert"
-          style={{
-            display: "flex",
-            justifyContent: "center",
-            color: "red",
-          }}
-        >
-          {errors.yourName.message}
-        </div>
-      )}
+      {errors.yourName && errMsg(errors.yourName)}
+
       <input
         className="Email-input"
         type="text"
@@ -91,18 +93,8 @@ const Email = () => {
       />
 
       <br />
-      {errors.email && (
-        <div
-          role="alert"
-          style={{
-            display: "flex",
-            justifyContent: "center",
-            color: "red",
-          }}
-        >
-          {errors.email.message}
-        </div>
-      )}
+      {errors.email && errMsg(errors.email)}
+
       <input
         className="Email-input"
         type="email"
@@ -117,18 +109,9 @@ const Email = () => {
       />
 
       <br />
-      {errors.myMessage && (
-        <div
-          role="alert"
-          style={{
-            display: "flex",
-            justifyContent: "center",
-            color: "red",
-          }}
-        >
-          {errors.myMessage.message}
-        </div>
-      )}
+
+      {errors.myMessage && errMsg(errors.myMessage)}
+
       <textarea
         className="Email-textarea"
         placeholder="Message"
