+++ FrontEnd/src/components/SpeakerPartnersComponent/SpeakerPartners.css
